requirements.txt
flask
langchain
langchain-openai
langchain-chroma
langchain-community
langchain-core
openai
python-dotenv